# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AgriVision AI – crop advisory platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plots:
    get:
      operationId: listPlots
      tags: [plots]
      summary: List all field plots
      responses:
        "200":
          description: List of plots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FieldPlot"
    post:
      operationId: createPlot
      tags: [plots]
      summary: Create a new field plot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldPlotInput"
      responses:
        "201":
          description: Created plot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldPlot"
  /plots/{id}:
    get:
      operationId: getPlot
      tags: [plots]
      summary: Get a specific plot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Field plot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldPlot"
        "404":
          description: Not found
    patch:
      operationId: updatePlot
      tags: [plots]
      summary: Update a field plot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldPlotUpdate"
      responses:
        "200":
          description: Updated plot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldPlot"
    delete:
      operationId: deletePlot
      tags: [plots]
      summary: Delete a field plot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Crop Diagnosis ───────────────────────────────────────────────────────
  /diagnosis/analyze:
    post:
      operationId: analyzeCrop
      tags: [diagnosis]
      summary: Analyze crop image for disease using Gemini AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosisInput"
      responses:
        "200":
          description: Diagnosis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CropDiagnosis"

  # ─── Soil Analysis ────────────────────────────────────────────────────────
  /soil/recommend:
    post:
      operationId: analyzeSoil
      tags: [soil]
      summary: Generate soil health and fertilizer recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoilTestInput"
      responses:
        "200":
          description: Soil analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoilAnalysisResult"

  # ─── Advisory Calendar ────────────────────────────────────────────────────
...
  /history/diagnoses:
    get:
      operationId: listDiagnoses
      tags: [history]
      summary: List historical crop diagnoses
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 20
        - name: plotId
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Paginated diagnoses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDiagnoses"
  /history/soil-tests:
    get:
      operationId: listSoilTests
      tags: [history]
      summary: List historical soil tests
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 20
        - name: plotId
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Paginated soil tests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSoilTests"
  /history/schedules:
    get:
      operationId: listSchedules
      tags: [history]
      summary: List historical advisory schedules
      parameters:
        - name: plotId
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdvisorySchedule"

  # ─── Weather Risk ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    FieldPlot:
      type: object
      required: [id, plotName, areaHectares, soilType, createdAt]
      properties:
        id:
          type: string
        farmerId:
          type: string
        plotName:
          type: string
        areaHectares:
          type: number
        soilType:
          type: string
        currentCrop:
          type: ["string", "null"]
        plantingDate:
          type: ["string", "null"]
        createdAt:
          type: string
    FieldPlotInput:
      type: object
      required: [plotName, areaHectares, soilType]
      properties:
        plotName:
          type: string
          minLength: 1
        areaHectares:
          type: number
          minimum: 0
        soilType:
          type: string
        currentCrop:
          type: string
        plantingDate:
          type: string
    FieldPlotUpdate:
      type: object
      properties:
        plotName:
          type: string
          minLength: 1
        areaHectares:
          type: number
          minimum: 0
        soilType:
          type: string
        currentCrop:
          type: string
        plantingDate:
          type: string

    # ─── Diagnosis ─────────────────────────────────────────────────────────
    DiagnosisInput:
      type: object
      required: [cropName, imageBase64, imageMimeType]
      properties:
        cropName:
          type: string
          minLength: 2
        imageBase64:
          type: string
          description: Base64-encoded image data
        imageMimeType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        observedSymptoms:
          type: string
        language:
          type: string
          enum: [en, hi, es, fr, sw]
        plotId:
          type: ["string", "null"]
    FertilizerStep:
      type: object
      required: [stage, fertilizerName, dosageKgPerHa, applicationMethod]
      properties:
        stage:
          type: string
        fertilizerName:
          type: string
        dosageKgPerHa:
          type: number
        applicationMethod:
          type: string
    CropDiagnosis:
      type: object
      required: [id, cropName, diseaseIdentified, confidenceScore, severityLevel, symptomsAnalysis, immediateAction, organicTreatment, chemicalTreatment, preventiveMeasures, createdAt]
      properties:
        id:
          type: string
        plotId:
          type: ["string", "null"]
        cropName:
          type: string
        diseaseIdentified:
          type: string
        scientificName:
          type: ["string", "null"]
        confidenceScore:
          type: number
        severityLevel:
          type: string
          enum: [Low, Moderate, Severe, Critical]
        symptomsAnalysis:
          type: array
          items:
            type: string
        immediateAction:
          type: string
        organicTreatment:
          type: array
          items:
            type: string
        chemicalTreatment:
          type: array
          items:
            type: string
        preventiveMeasures:
          type: array
          items:
            type: string
        localizedNote:
          type: ["string", "null"]
        imageUrl:
          type: string
        createdAt:
          type: string

    # ─── Soil Analysis ─────────────────────────────────────────────────────
    SoilTestInput:
      type: object
      required: [phLevel, nitrogenPpm, phosphorusPpm, potassiumPpm, targetCrop]
      properties:
        plotId:
          type: ["string", "null"]
        phLevel:
          type: number
          minimum: 3.0
          maximum: 11.0
        nitrogenPpm:
          type: number
          minimum: 0
          maximum: 1000
        phosphorusPpm:
          type: number
          minimum: 0
          maximum: 1000
        potassiumPpm:
          type: number
          minimum: 0
          maximum: 1000
        organicMatterPct:
          type: number
          minimum: 0
          maximum: 20
        targetCrop:
          type: string
          minLength: 2
        language:
          type: string
          enum: [en, hi, es, fr, sw]
    SoilAnalysisResult:
      type: object
      required: [id, overallHealthRating, phAnalysis, nutrientDeficiencies, organicAmendments, fertilizerRegimen, irrigationAdvice, createdAt]
      properties:
        id:
          type: string
        plotId:
          type: ["string", "null"]
        overallHealthRating:
          type: string
          enum: [Optimal, Deficient, "Toxic/Excess", Imbalanced]
        phAnalysis:
          type: string
        nutrientDeficiencies:
          type: array
          items:
            type: string
        organicAmendments:
          type: array
          items:
            type: string
        fertilizerRegimen:
          type: array
          items:
            $ref: "#/components/schemas/FertilizerStep"
        irrigationAdvice:
          type: string
        createdAt:
          type: string

    # ─── Advisory Schedule ─────────────────────────────────────────────────
...
    WeeklyTask:
      type: object
      required: [week, startDate, tasks, category]
      properties:
        week:
          type: number
        startDate:
          type: string
        tasks:
          type: array
          items:
            type: string
        category:
          type: string
        notes:
          type: ["string", "null"]
    AdvisorySchedule:
      type: object
      required: [id, cropType, startDate, expectedHarvestDate, weeklyTasks, totalDays, createdAt]
      properties:
        id:
          type: string
        plotId:
          type: ["string", "null"]
        cropType:
          type: string
        startDate:
          type: string
        expectedHarvestDate:
          type: string
        weeklyTasks:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyTask"
        totalDays:
          type: number
        createdAt:
          type: string

    # ─── Weather Risk ──────────────────────────────────────────────────────
...
    PaginatedDiagnoses:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CropDiagnosis"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
    PaginatedSoilTests:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SoilAnalysisResult"
        total:
          type: number
        page:
          type: number
        limit:
          type: number