# lib/api-zod/src/generated/types/advisorySchedule.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AgriVision AI – crop advisory platform API
 * OpenAPI spec version: 0.1.0
 */
import type { WeeklyTask } from './weeklyTask';

export interface AdvisorySchedule {
  id: string;
  /** @nullable */
  plotId?: string | null;
  cropType: string;
  startDate: string;
  expectedHarvestDate: string;
  weeklyTasks: WeeklyTask[];
  totalDays: number;
  createdAt: string;
}