# lib/api-zod/src/generated/types/cropDiagnosis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AgriVision AI – crop advisory platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CropDiagnosisSeverityLevel } from './cropDiagnosisSeverityLevel';

export interface CropDiagnosis {
  id: string;
  /** @nullable */
  plotId?: string | null;
  cropName: string;
  diseaseIdentified: string;
  /** @nullable */
  scientificName?: string | null;
  confidenceScore: number;
  severityLevel: CropDiagnosisSeverityLevel;
  symptomsAnalysis: string[];
  immediateAction: string;
  organicTreatment: string[];
  chemicalTreatment: string[];
  preventiveMeasures: string[];
  /** @nullable */
  localizedNote?: string | null;
  imageUrl?: string;
  createdAt: string;
}