# lib/api-zod/src/generated/types/paginatedSoilTests.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AgriVision AI – crop advisory platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SoilAnalysisResult } from './soilAnalysisResult';

export interface PaginatedSoilTests {
  items: SoilAnalysisResult[];
  total: number;
  page: number;
  limit: number;
}